ADD: launch for vision_bringup
