+++ realsense2_camera/launch/rs_kiwi.launch.py
@@ -1,0 +1,72 @@
+"""Launch the vision stack in a component container."""
+import os
+import cv2
+
+from ament_index_python.packages import get_package_share_directory
+from launch_ros.actions import LoadComposableNodes
+from launch.actions import DeclareLaunchArgument, IncludeLaunchDescription, GroupAction
+from launch.conditions import IfCondition, UnlessCondition
+from launch_ros.descriptions import ComposableNode
+from launch.substitutions import LaunchConfiguration
+from launch import LaunchDescription
+from launch.launch_description_sources import PythonLaunchDescriptionSource
+
+
+def generate_launch_description():
+
+    use_respawn = LaunchConfiguration("use_respawn")
+
+    vision_config = os.path.join(
+        get_package_share_directory("vision_bringup"), "launch", "vision_params.yaml"
+    )
+
+    return LaunchDescription(
+        [
+            # -------------- COMPOSITION -------------------------------
+            DeclareLaunchArgument(
+                "use_composition",
+                default_value="False",
+                description="Whether to use node composition",
+            ),
+            DeclareLaunchArgument(
+                "use_respawn",
+                default_value="False",
+                description="Whether to respawn if a node crashes. Applied when composition is disabled.",
+            ),
+            GroupAction(
+                condition=IfCondition(LaunchConfiguration("use_composition")),
+                actions=[
+                    LoadComposableNodes(
+                        target_container="vision_kronos",
+                        composable_node_descriptions=[
+                            ComposableNode(
+                                parameters=[vision_config],
+                                package="realsense2_camera",
+                                plugin="realsense2_camera::RealSenseNodeFactory",
+                                name="camera",
+                                namespace="camera",
+                                extra_arguments=[{"use_intra_process_comms": True}],
+                            )
+                        ],
+                    ),
+                ],
+            ),
+            # -------------- NO COMPOSITION ----------------------------
+            GroupAction(
+                condition=UnlessCondition(LaunchConfiguration("use_composition")),
+                actions=[
+                    IncludeLaunchDescription(
+                        PythonLaunchDescriptionSource(
+                            [
+                                os.path.join(
+                                    get_package_share_directory("realsense2_camera"),
+                                    "launch",
+                                ),
+                                "/rs_launch.py",
+                            ]
+                        )
+                    )
+                ],
+            ),
+        ]
+    )
